netlink: consistently use uint16_t for family id
Reviewed by:		melifaro
Differential Revision:	https://reviews.freebsd.org/D47547

+++ sys/netlink/netlink_domain.c
@@ -256,7 +256,7 @@ nl_send_group(struct nl_writer *nw)
 }
 
 bool
-nl_has_listeners(int netlink_family, uint32_t groups_mask)
+nl_has_listeners(uint16_t netlink_family, uint32_t groups_mask)
 {
 	return (V_nl_ctl != NULL);
 }
